+++ malachite-test/tests/integer/arithmetic/div_exact_signed_limb.rs
@@ -36,7 +36,7 @@ fn test_div_exact_signed_limb() {
 
         let q = Integer::from_str(i)
             .unwrap()
-            .div_round(j, RoundingMode::Exact);
+            .div_round(Integer::from(j), RoundingMode::Exact);
         assert_eq!(q.to_string(), quotient);
 
         let q = rug_div_exact_signed_limb(rug::Integer::from_str(i).unwrap(), j);
@@ -246,7 +246,7 @@ fn div_exact_signed_limb_properties_helper(n: &Integer, i: SignedLimb) {
     assert!(quotient_alt.is_valid());
     assert_eq!(quotient_alt, quotient);
 
-    let quotient_alt = n.div_round(i, RoundingMode::Exact);
+    let quotient_alt = n.div_round(Integer::from(i), RoundingMode::Exact);
     assert!(quotient_alt.is_valid());
     assert_eq!(quotient_alt, quotient);
 
@@ -281,7 +281,7 @@ fn div_exact_signed_limb_properties() {
             assert!(quotient_alt.is_valid());
             assert_eq!(quotient_alt, quotient);
 
-            assert_eq!(i.div_round(n, RoundingMode::Exact), quotient);
+            assert_eq!(Integer::from(i).div_round(n, RoundingMode::Exact), quotient);
 
             assert_eq!(i.div_exact(-n), -&quotient);
